Fixed grabs not reducing blood loss on grabbed body parts (#2751)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This PR fixes an issue which causes grabbing a bleeding body part to not
reduce the blood loss on that body part.

A body part's blood loss was intended to be reduced by 50% if that body
part was grabbed - and the remainder by a further 50% for every
additional grab (so, two grabs would be a reduction of 75%, three would
be 87.5%, and so on). This code already existed, and in fact wasn't
modified in this PR.

However, the code dealing with _grabs_ would fail to add the grab to the
limb's "grabbedby" list. As a consequence, the code described above was
essentially dead code - the condition for the blood loss reduction would
never be fulfilled under any circumstances. This is what the PR
modifies, by adding the grab to the body part's grabbedby list (and
thereafter removing it at the appropriate time).

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

This is something that was always intended to work, and would be
valuable in buying a little more time while taking heavily-wounded
characters to the church or a physician, without actually doing anything
to treat the wounds on its own. This is particularly valuable in light
of the recent limits placed on character revival - a little more time to
get somebody treated could go a long way.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

It's a little tricky to show this in a way that's unambiguous as to the
reduction in blood loss since, as far as I'm aware, there's no display
that shows an actual numerical amount in-game. Likewise, I could do so
much damage that I bled out rapidly, but this would mean I'd be doing a
lot of prior damage to myself, meaning a small health pool to show the
effects from. So, instead, I got in my test environment, wounded myself
a bunch, and then compared the bleeding in the debugger before and after
grabbing the affected body part, without otherwise modifying anything.

The wounds:

![image](https://github.com/user-attachments/assets/eb2d10e3-cb40-49b7-80fc-0e1a554310da)

Before grab, 21 blood loss:

![image](https://github.com/user-attachments/assets/39f17582-378e-4c60-9537-7ad8365646ca)

After grab, 10.5 blood loss:

![image](https://github.com/user-attachments/assets/7665e30e-0000-4c6a-9ce7-791f1a4637c6)

I've additionally ensured that grabbing mobs still works correctly, as
well as objects becoming embedded mid-combat.

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->