Dead examine message fix (#2731)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Prevents "[someone] looks at [something]" message in chat when [someone]
is dead
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Fix
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/8291910c-adc2-4f68-9b12-9ca7867c4818)

![image](https://github.com/user-attachments/assets/23f33a91-c76f-4726-b592-095ac8261797)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->