Adds a new spell: enlarge (#2674)
## About The Pull Request

Inspired by the enlarge spell in Azure Keep
(https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/1325)
It's been ported over to our codebase using the buff system instead.
There is a few implementation differences due to that.

Currently, there is a minor problem where the sprite does not update
until there is anything that calls the update function, it seems that
owner.update_transform() does not work as intended and I don't know how
to fix it.
In essence what this means is that you have to take a single step or
movement for the visual part of the spell to take effect.

What it does:
Makes you bigger like the giant trait
Adjusts your stats +2str+2con-2spd
And because it's funny, it gives you the bedbreaker trait.
It costs 2 points and gives a total of +2 statpoints, it's balance is
closely tied to the haste spell from which I took the majority of
numbers. Haste costs two points and gives you +3spd for the same
duration and cooldown.

## Why It's Good For The Game

It's a cool spell that has both combat and out of combat uses. Just some
fun content. I plan to perhaps port a few more spells I've seen in
azurekeep such as longstrider, counterspell and lesser knock(depending
on where tes lockpick pr goes).


## Proof of Testing (Required)
Done on local server, PLEASE DOUBLECHECK THIS IS MY FIRST TIME DOING
ANYTHING BEYOND MINOR NUMBER TWEAKS.
Enlarging myself. The ladies will love this one.

![enlarge1](https://github.com/user-attachments/assets/23217bc5-32a7-4713-8a69-64cac5be853f)
Enlarging a goblin that's been chasing me. He is slower now but I feel
like I've still made the problem worse.

![enlarge2](https://github.com/user-attachments/assets/051a38ac-e689-4e8b-803b-29b01e09515d)