Removes Magos Refugee Patron Lock (#2750)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Allows magos refugees to have either Noc or Zizo as their patron.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Magos refugees are forced into only Noc (except for Battle Magos for
some reason). This gives more opportunities for different flavors of
magic users from afar. I made sure to check if there was a specific
reason as to why they were locked only to Noc, but it doesn't seem like
there is one.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![Screenshot 2025-04-18
172617](https://github.com/user-attachments/assets/56794a03-50b1-468f-8e8d-80f031ca8629)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->