Adds Lupian-Only Noises (#2752)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

- Adds four new common canine noises to lupian characters: Howl, Growl,
Whine, and Bark.
- All sounds are from https://pixabay.com/ , are royalty free, and are
edited/cut by me. The growl sound is one of the volf sounds, but it fits
well enough anyways.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

- Tabaxi get unique sounds, why shouldn't lupians?
- The dogs can now bark back at the stupid hissy cats.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)



https://github.com/user-attachments/assets/3cbb2381-a778-4615-9ca5-6dc11cc1f9f9



<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->