PP Improvements (#2754)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

It pretty much substitutes the New Player Panel (NPP) with the Super
Player Panel (SPP)
(I won't call it New New, or call the other one 'old' because of an OLD
PP existing.)

The SPP will substitute the NPP.

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Removed broken and crashing functions of the New Player Panel with a
SUPER Player Panel
Still have plans to add more to it, but as of now it's way better than
the New Player Panel.

## Proof of Testing (Required)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->

That's the SUPER PP. Say hi!

![image](https://github.com/user-attachments/assets/947d5f00-205c-48ee-8ee5-4fa50fbab064)

[The right side options open up too]

![image](https://github.com/user-attachments/assets/30b630c4-749e-40c4-9bca-04b1b542062b)

![image](https://github.com/user-attachments/assets/6d7557cd-07a5-4973-bc06-f00c1561b790)

![image](https://github.com/user-attachments/assets/a8c59916-0e2f-4145-b051-713ab778d187)

---------

Co-authored-by: Useroth <37159550+Useroth@users.noreply.github.com>