Proper directions for sounds (#2758)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Instead of calling get_dir to... get direction... of sound, which would
get you this

![image](https://github.com/user-attachments/assets/446ca0ce-77fa-4215-bd10-5df56243f853)
gets a direction based on the angle between user and source, and then
converts it into direction, which gives this

![image](https://github.com/user-attachments/assets/d95a2a5d-39c2-4de1-9232-d69f976b84c1)


Changed that for speaking/screaming (the arrows (⇗) thing) and signal
horn. If there are any other things similar to those which needs to be
changed, let me know, because i cant remember any more of them.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Proper directions of sounds
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/599fb8ce-267a-480a-b367-11400a03d583)

![image](https://github.com/user-attachments/assets/c944e888-41bf-4e14-a055-d54f33819180)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->