Adds Necra censer, and smoke effect. (#2737)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Adds the following:

- Necra censer, an item for morticians to clean the town of blood
- Adds the item to their backpack
- Gives it a new sound
- Attack self (click on item) to use
- makes new Necra censer smoke that cleans floors (finally utilizing the
code that fixes smoke from not moving!)
- Adds sprite for censer

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The town gets very dirty very fast and cleaning it manually is hell.
There's no knacker role and morticians are the closest thing we have to
cleaners when they aren't burying the dead. This allows them to clean
several large patches in a relatively short amount of time (It's still a
3 second do_after).

This will also help alleviate any pain if blood in the future is going
to cause any diseases or sickness.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/9b1df785-cd1a-4e11-8648-f3babfdba89f


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->