Bandit Balance and tweaks (#2707)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR puts all bandit classes in line with each other stat wise; While
dropping bandit skills down to journeyman.
It also fixes outlier issues like some classes having low swimming or
athletics. or classes like Knave having extra stat points/extra
wrestling skill/expert knives.

Subsequently, It also removes the 'wanted' opt in with higher stat and
skills.
Instead, all bandits are made wanted; However You can only tell if a
person is wanted if you've looked at a bandit wanted poster in the past
20 minutes; or if you're a leadership role such as mayor, knight
captain, warden, or sheriff.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
This PR is intended to pair with the bandits receiving a bit of a hike
in power with the Hoardmaster kidnapping PR, allowing them access to
steel weaponry. The design goal of bandits has always been a minor
antagonist; Due to recent map changes, combined with newer content this
is no longer the case.
Bringing bandits back down to level 3 (in most cases) is a necessary
thing unfortunately.

Furthermore; Finding a decent 'alternative' to reward for being wanted
was- difficult. It could have supplied extra favor; But in truth, most
people almost always went wanted for the extra skill in fighting. And
not having bandits with the wanted status proved dentrimental to the
server- It lead to bandits usually not getting into fights and just
walking past guards into town.

I present an alternative utilizing wanted posters.
Examining a wanted poster provides about an in game days worth of time
for anyone to identify a bandit. This means if a guard hasn't looked at
a poster and finds someone suspicious they're more then welcome to haul
that person to a poster and check.
Eventually I'll see about making it so wanted posters can be torn down
after a delay.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/78e9df07-7d60-430f-981f-f15329130746)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->