Adds chitin to wildkin skintype descriptor (#2790)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Spider and insect like character have to compromise with descriptors
that don't fit.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Just a neat little thing that adds depth.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![435275449-2343b5a6-d917-471f-b97b-ed764b7dcc90](https://github.com/user-attachments/assets/09f67282-5962-4cc9-b0fa-15dd34ee7451)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->