updates carp, clownfish, eel, and angler sprites (#2793)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Changes the sprites for clownfish, angler, carp, and eel: example
comparison here (not actual layout of file)

![image](https://github.com/user-attachments/assets/12855abe-8b0f-4da0-aa96-8d0f0a807eb4)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Makes them a little bit prettier. 
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
just a dmi change but it compiles and I double checked to make sure I
had no empty sprites.
<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->