Lock the kitchen doors to a kitchen-specific key (#2797)
## About The Pull Request
Changes the lock for the manor's kitchen to a kitchen-specific key.
Adds two spare kitchen keys to the kitchen to go along with the two
spare manor keys.
Gives nobles and kitchen staff the kitchen key.

## Why It's Good For The Game
Since #2976 was rejected by Alanii, I suppose this is the only option we
have left to fix the retinue breaking into the kitchen to stuff
themselves with raw ingredients *while the maids are cooking.*

## Proof of Testing (Required)
TBD.