Wenchmaxxing - Whitevein Role Buffs (#2746)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Some expansion of and changes to the Nightmaster and Prostitute role.
Renames Prostitute to "Nightswain/Nightmaiden" because it's a way cooler
name.

**Nightmaster Changes**
The Nightmaster is losing henchmen, so needs to be able to better deal
with problems that come up. He also lacks some skills that make sense
for his position. He also was not getting roundstart money, and needs it
to hire guards or to make early round purchases. Besides, he's a noble
and a business owner: Why is he so broke?!

- Nightmaster gains +2 endurance
- Nightmaster gains 2 levels in athletics
- Nightmaster gains whips and flails 2, why didn't he already have this?
- Nightmaster gains music 2 to reflect his noble upbringing
- Nightmaster now starts with 69 mammons. This is less than the cook
and/or bartender.
- Nightmatrons no longer spawn with the silk coat. It does not have a
female sprite and they would appear topless.

**Prostitute Changes**
The role has been renamed to Nightswain/Nightmaiden so it would be a
little less on the nose. Now has three subclasses to represent different
wench archetypes.

The subclasses are as follows.

Harlot: Closest to old prostitute. Streetwise wenches with sticky
fingers. They have sneaking and stealing skills, and have knives 2 and
nutcracker to make them surprisingly mean if the average towner tries to
mess with one. Starts with a corset to hide things in.

Coquette: The classy girl, the expensive one. Knows how to read, better
at playing music, even has some minor domestic skills like low level
sewing and cooking. They also start with a dress that actually covers
everything.

Tickler: Leather Zaddy/Dominatrix. Higher strength, knows their way
around a whip and knows how to manhandle a submissive. Doesn't have a
lot of other skills to speak of. Is not very bright. Spawns with a
leather bikini, or leather pants and no shirt, along with a whip. Only
one allowed at a time - I erred on the side of caution with this one.


**Fabled Lover Buff**

Fabled lover has been changed to apply a long-lasting +2 fortune buff in
addition to giving triumph. This buff can only be obtained once per
round per person.

**Sexcon Fixes**

The sexcon is now more reliable about applying the effects of Fabled
Lover in the intended way. It also no longer runtimes when someone
climaxes due to a sex action initiated by the receiving partner. I have
also taken the liberty of having it add the "sexjuice" reagent to
characters receiving an ejaculation. You're welcome.

**Corset Knives**

You can now hide small objects inside corsets. Right clicking the corset
with an empty hand will remove the object without making a sound or
producing a message in chat.


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

It's an underplayed role. I wanted to give it some support for playing a
larger variety of character archetypes, and to make it a little more fun
for people who actually want to play these characters for more than just
ERP.

Nightmaster needed a bit of a tune up. The role felt incomplete and was
missing some skills that I felt would fit him quite well. The added
money at roundstart both makes sense for a business owner and drug
dealer and will be necessary for hiring guards going forward, as well as
allowing some potential to actually do more in a round. The Silk Coat is
broken on females.

Corset knives was just a random idea I had and decided to code it. It's
super fun and cool though!

Fabled Lover buff is to give a mechanical incentive to visit the
Whitevein, particularly early round. Triumphs are a largely deprecated
system. The people who would be interested in the buff are going to be
people who have money to spend on it, and this will incentivize them to
go and spend money earlier in the round instead of the typical Whitevein
experience of people showing up near round end to show off their 1500
mammon hoard that will never actually be used because the round is
ending. Money flowing is a good thing, giving one of the town's services
something actually worth paying for is a great thing.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


![image](https://github.com/user-attachments/assets/81ba6bc9-fb40-47ee-8e5d-de77b0578fd6)

![image](https://github.com/user-attachments/assets/f2302f0f-c0a9-4a65-9d0e-e68e9ef60233)

![image](https://github.com/user-attachments/assets/b0765990-0ab6-46c0-889d-42f37ba97261)

![image](https://github.com/user-attachments/assets/ab774df7-cfb4-4299-9330-c575043ec341)

![image](https://github.com/user-attachments/assets/bbc24241-405e-4fc0-b0ae-d9102142bc73)




<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->