Allows Create Item interface to search by item name, as well as type (#2770)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR modifies the Game Panel's Create Item and Quick Create Item
interfaces to allow searching for items by in-game name, as well as by
object type path. Two radio buttons have been added to the interface to
allow toggling between the old and new behaviour, with the new behaviour
(search by in-game name) being the default.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Spawning items is a function admins may occasionally need to do to aid
players - helping recover from a bug that consumed an item, for example,
or even in a more Game Master-like sense by creating an item in response
to a prayer, or for an event. Most players won't interact with this
functionality in the first place, but for those they do, being able to
search for an item by its more familiar in-game name rather than its
type path is convenient, particularly in cases where the names differ
significantly between the two.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
The search results for the new, search-by-in-game-name functionality:

![image](https://github.com/user-attachments/assets/d8d0c5b4-d33b-4832-abd1-15729e1c24d3)

The item having successfully been spawned after being searched for by
name:

![image](https://github.com/user-attachments/assets/f76d646c-4414-404e-b3d1-bac9a6b9ef43)

The same item being searched for by its type in code:

![image](https://github.com/user-attachments/assets/bfc392fb-8e15-4da7-aed9-63abb812a784)

And the same item, having been spawned successfully after being searched
for by tpye:

![image](https://github.com/user-attachments/assets/2ad5c973-75ff-4390-926e-a4fe4b996c54)

I've additionally tested both the Create Object and Quick Create Object
interfaces with these changes, as well as testing spawning a number of
items in both modes - but I don't want to drown this PR with dozens of
screenshots.

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->