Ports Azure's text formatting. (#2771)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Port's Free's text formatting over to here
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/1720

**HOW TO USE:**

- Requires a start and end example: `This is +bold text+` will output:
This is **bold text**
- bold text symbol: `+`
- Italics text symbol: `|`
- underline text symbol: `_`

Note: The only issue I've seen (and Free also mentioned this) is that
you cannot start a sentence with a bold word because it conflicts with
using + for whispers. We _could_ change whisper to # but would need to
make it very clear for people until they get used to this or find some
way to check a whole block of text for a second + in some edge case. We
could also change the bold char symbol but I'd rather keep it consistent
between the servers for clarity sake (and because I play on both and
would go mad trying to remember which is which)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Fromatting is fun.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/e36e0279-c323-4892-9aa5-ecf77ea34635)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->