Title friendly marriage (#2788)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Makes marriage name changing title friendly
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Currently renaming of characters during marriage is fucked up, because
this system was created before we got titles in names, so it was no
friendly to space between title and rest of the name, resulting in usage
of first name in remaning, instead of last name
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/3ad78464-3cfb-4462-8d47-9588bcee65a0)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->