Gives 2 dagger knaves lockpicks (#2811)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Gives the 2 dagger knave subclass lockpicks
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
They're the most hidden 'thief' type of bandits but never spawned with
lockpicks.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/ac274dd5-53ec-40dc-ac94-4287a30b6527)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->