Reduce the door bump cooldown (#2817)
## About The Pull Request
Reduces the 6 second cooldown between interacting with doors to a 0.3
second cooldown. Each door still has a 2 second bump cooldown, but mobs
can do approximately three doors per second, which is enough to stop it
from letting you charge through a series of doors at full speed or
whatever.

## Why It's Good For The Game
This is how base TG has it and the *6 second fucking cooldown* is
cancerous.

## Proof of Testing (Required)
This was in #2794 and is being split out.