Framework and Basic Loop
More content etc. forthcoming.

Adds the new Vars to Rituals (Cost and Difficulty)
Adds the New Skill and adds it to Zizo Dualists for testing
Adds the old cult verbs to Zizo Dualist
Adds a new Var to mind (zizofavor)
Adds a few new Rituals and rebalances the old ones