(AP port) voice pack selection (#2787)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Ports some of the features from
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/commit/cfb127f53ed9c4d658e2d4ca91d44bcb728a31eb

To allow people to choose if they wish to have a masculine or feminine
voice. I plan on expanding this some time in the future to allow for
other voice packs to get used (e.g. if you want dwarf or goblin voices)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
People want their femboys to sound girly... Plus woman with a manly
voice is funny. It also lays the ground work out for me to expand on it.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
Didn't see any runtimes and I don't think there could be any.


https://github.com/user-attachments/assets/b8753cb7-3b7c-401c-a742-9595a6581648


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->