fixes silk glove on mob sprites (#2838)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
see title
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Reported as broken:

![image](https://github.com/user-attachments/assets/29080a88-f923-48c0-9cb4-cf6ee1b7ff1e)

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/d38133a3-1a5f-4b4c-87a0-08608e64c53e)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->