removes prob roll to make jesters white (#2830)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Removes the grenzel prob roll for jester.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
If you're a non grenzelhoft northener your character will be white but
your headshot might show something completely different.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
haven't tested it but if just commenting out an if statement like this
breaks anything I'd be shocked.
<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->