Fixes issues with Cure Rot (#2850)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR fixes two issues with Cure Rot:
- Cure Rot killing targets who have the deadite infection, but who have
not yet become deadites.
- Cure Rot applying the Rot-Touched condition to targets who aren't
deadites and haven't rotted (including uninfected targets!)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
In gameplay, Cure Rot is intended to be the most potent and reliable way
of dealing with the rot and associated deadite infection. However, at
present, Cure Rot will reliably kill a target who is merely infected -
as well as applying the Rot-Touched debuff, preventing the target from
ever being cured of the rot again. Additionally, if Cure Rot is ever
incorrectly targeted - which isn't difficult to do when an Unknown is
being restrained by another Unknown - it will apply the Rot-Touched
debuff to the incorrect target, despite there being no rot to remove.

This fix improves the situation by adding an incentive to treat the
infection: treating it in time will save not only the target's life, but
spare them the rot-touched condition. Additionally, Cure Rot now can't
be used to maliciously give somebody the rot-touched debuff, and won't
apply it in cases of accidental application; putting the spell on
cooldown for a while is already punishment enough for targeting the
wrong person, in my opinion.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
A character is infected with the deadite infection, given an Eoran bud,
and allowed to turn.

![image](https://github.com/user-attachments/assets/aef24d09-238f-476f-94b9-401f50b10efa)

The deadite rises.

![image](https://github.com/user-attachments/assets/00382d35-cdfb-4bea-b776-93418c0239e3)

I use Cure Rot, correctly killing the deadite, and leaving it
rot-touched.

![image](https://github.com/user-attachments/assets/5066d59a-74ab-4b0e-b61e-c902ac290391)

I apply the infection to a new character.

![image](https://github.com/user-attachments/assets/9ed6a11d-e177-493b-b510-661b3fe89132)

I cure the infection, leaving the target alive and free of the
rot-touched condition as a reward for my swift action.

![image](https://github.com/user-attachments/assets/4619d4f1-c9db-4319-8226-59ebe0f090e2)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->