Fix for deadite infection causing uncontrollable biting (#2845)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR fixes an issue which causes characters infected with the deadite
infection to bite those around them uncontrollably.

This issue is caused by the zombie antag datum's on_life proc not
considering whether or not the character had actually turned into a
deadite yet - hence, anybody infected would immediately begin biting
anybody nearby, rather than only doing so after actually becoming a
deadite.

The fix is a simple check against the has_turned variable before
proceeding to attempt to bite anything.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
This corrects the infection to its intended behaviour - the victim is a
danger because they'll soon turn into a deadite, but still retains
control of themselves for the time being. The uncontrollable biting was
unintended behaviour which didn't fit well with the player otherwise
retaining full control of themselves, and additionally made further
interaction with them difficult.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
Myself, infected and not biting the character next to me.
(force_apply_deadite_bite is a debug proc written specifically to test
this and similar issues, and isn't included in the PR.)

![image](https://github.com/user-attachments/assets/84f40b95-a6de-4ba4-bf8b-48bc05f50ee5)

And indeed after turning, I still assault the character as expected.

![image](https://github.com/user-attachments/assets/9726e65e-c599-4cd4-9ab9-e86f79374cd0)


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->