Returns old court magos +1 to arcane magic (#2843)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
brings it back. Seemed to get deleted when someone did not understand
what the proc did?
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
He's supposed to have it. At least that's what a PR from before
indicates and it never got fixed.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
before: 

![image](https://github.com/user-attachments/assets/37846c7b-83ac-41e6-ad19-2a266c8a676e)
with PR:

![image](https://github.com/user-attachments/assets/8a69debd-2484-4834-90af-e0e7fc77f00e)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->