Datumizes stew code. (#2779)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Does the following:
- makes pots 66oz (six bowls, requires 6 items, more on that later)
- new datum system for cooking stew recipes

The new data for the pot recipes makes it easy to edit and manipulate.
You can also add non food items to create reagent types or which ever
you'd like in the future.
The example here is `/datum/pot_recipe/poo` where it accepts an
`/obj/item/natural` item

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Stew has been broken in weird ways for a while. This allows people to
make a filling meal, albeit not exactly great, by making stew once more
with a straight forward system of 1 food item per bowl of stew. This is
rather in line with what other food currently does due to a lot of it
mostly existing to fill stomachs without giving any buffs.

This also allows us to try adding other things into stew that aren't
traditionally food easily and may also allow us to make use of the
`end_cooking` proc to perhaps spawn items on creation at the pot's
location (it could be how we make spaghetti or rice if we ever add stuff
like that.)
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
boil check

![image](https://github.com/user-attachments/assets/97bb8df9-5aff-44d7-84bc-bc43da938383)

water check

![image](https://github.com/user-attachments/assets/88e4e075-ebdc-41a3-b08e-873eba4b38ee)

6 bowls from one pot

![image](https://github.com/user-attachments/assets/d8cb8e7c-834c-4d0d-8bd6-af5467c47c90)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->